+++ tests/cibyl/unit/sources/elasticsearch/test_api.py
@@ -19,9 +19,8 @@ from unittest import TestCase
 from unittest.mock import Mock, patch
 
 from cibyl.cli.argument import Argument
-from cibyl.exceptions.source import InvalidArgument, MissingArgument
+from cibyl.exceptions.source import MissingArgument
 from cibyl.sources.elasticsearch.api import ElasticSearch
-from tests.cibyl.utils import OpenstackPluginWithJobSystem
 
 
 class TestElasticSearch(TestCase):
@@ -408,227 +407,3 @@ class TestElasticSearch(TestCase):
         es_api.ensure_source_setup()
         self.assertTrue(es_api.is_setup())
         mock_client.assert_called_with("https://example.com", 9200)
-
-
-class TestElasticSearchOpenstackPlugin(OpenstackPluginWithJobSystem):
-    """Test cases for :class:`ElasticSearch` with openstack plugin."""
-
-    def setUp(self) -> None:
-        self.es_api = ElasticSearch(elastic_client=Mock())
-        self.job_hits = [
-                    {
-                        '_id': 1,
-                        '_score': 1.0,
-                        '_source': {
-                            'job_name': 'test',
-                            'job_url': 'http://domain.tld/test',
-
-                        }
-                    },
-                    {
-                        '_id': 2,
-                        '_score': 1.0,
-                        '_source': {
-                            'job_name': 'test2',
-                            'job_url': 'http://domain.tld/test2',
-                        }
-                    },
-                    {
-                        '_id': 3,
-                        '_score': 1.0,
-                        '_source': {
-                            'job_name': 'test3',
-                            'job_url': 'http://domain.tld/test3',
-                        }
-                    },
-                    {
-                        '_id': 4,
-                        '_score': 1.0,
-                        '_source': {
-                            'job_name': 'test4',
-                            'job_url': 'http://domain.tld/test4',
-                        }
-                    }
-        ]
-
-        self.build_hits = [
-                    {
-                        '_source': {
-                            'job_name': 'test',
-                            'job_url': 'http://domain.tld/test/',
-                            'build_result': 'SUCCESS',
-                            'build_id': '1',
-                            'time_duration': 20,
-                            'build_url': 'http://domain.tld/test/7',
-                            'ip_version': 'ipv4',
-                            'network_backend': 'local_area_n',
-                            'test_name': 'it_is_just_a_test',
-                            'test_time': '0.001',
-                            'test_class_name': 'folder.file.ClassName'
-                        }
-                    },
-                    {
-                        '_source': {
-                            'job_name': 'test2',
-                            'job_url': 'http://domain.tld/test2/',
-                            'build_result': 'FAIL',
-                            'build_id': '2',
-                            'time_duration': 10,
-                            'build_url': 'http://domain.tld/test2/8',
-                            'ip_version': 'ipv6',
-                            'network_backend': 'local_area_n',
-                            'test_name': 'it_is_just_a_test2',
-                            'test_time': '0.0001_bad_parsed',
-                        }
-                    }
-        ]
-
-        # This is an aggregation + query results
-        self.tests_hits = [
-            {
-                '_source': {
-                    'job_name': 'test',
-                    'job_url': 'http://domain.tld/test/'
-                },
-                'key': 'test',
-                'last_build': {
-                    'hits': {
-                        'hits': [
-                            {
-
-                                '_source': {
-                                    'job_name': 'test',
-                                    'job_url': 'http://domain.tld/test/',
-                                    'test_setup': 'rpm',
-                                    'ip_version': 'ipv4',
-                                    'test_suites': 'designate,neutron,octavia'
-                                }
-                            }
-                        ]
-                    }
-                }
-            },
-            {
-                '_source': {
-                    'job_name': 'test2',
-                    'job_url': 'http://domain.tld/test2/'
-                },
-                'key': 'test2',
-                'last_build': {
-                    'hits': {
-                        'hits': [
-                            {
-                                '_source': {
-                                    'job_name': 'test2',
-                                    'job_url': 'http://domain.tld/test2/',
-                                    'ip_version': 'ipv4',
-                                }
-                            }
-                        ]
-                    }
-                }
-            }
-        ]
-
-    @patch.object(ElasticSearch, '_ElasticSearch__query_get_hits')
-    def test_get_deployment(self: object, mock_query_hits: object) -> None:
-        """Tests that the internal logic from
-        :meth:`ElasticSearch.get_deployment` is correct.
-        """
-        mock_query_hits.return_value = self.tests_hits
-
-        jobs_argument = Mock()
-        jobs_argument.value = ['test']
-
-        ip_address_kwargs = Mock()
-        ip_address_kwargs.value = []
-
-        jobs = self.es_api.get_deployment(jobs=jobs_argument,
-                                          ip_version=ip_address_kwargs)
-        deployment = jobs['test'].deployment.value
-        network = deployment.network.value
-        self.assertEqual(network.ip_version.value, '4')
-        self.assertEqual(deployment.topology.value, '')
-
-    def test_spec_deployment(self: object):
-        self.es_api.get_jobs = Mock(side_effect=self.job_hits)
-        spec = Mock()
-        spec.value = []
-
-        with self.assertRaises(InvalidArgument):
-            self.es_api.get_deployment(spec=spec)
-
-    @patch.object(ElasticSearch, '_ElasticSearch__query_get_hits')
-    def test_deployment_filtering(self: object,
-                                  mock_query_hits: object) -> None:
-        """Tests that the internal logic from
-        :meth:`ElasticSearch.get_deployment`
-            is correct.
-        """
-        mock_query_hits.return_value = self.tests_hits
-
-        jobs_argument = Mock()
-        jobs_argument.value = ['test']
-
-        # We need to mock the Argument kwargs passed. In this case
-        # ip_address
-        ip_address_kwargs = Mock()
-        ip_address_kwargs.value = ['4']
-
-        builds = self.es_api.get_deployment(jobs=jobs_argument,
-                                            ip_version=ip_address_kwargs)
-
-        deployment = builds['test'].deployment.value
-        network = deployment.network.value
-        self.assertEqual(network.ip_version.value, '4')
-        self.assertEqual(deployment.topology.value, '')
-
-    @patch.object(ElasticSearch, '_ElasticSearch__query_get_hits')
-    def test_get_deployment_filter_test_setup(self, mock_query_hits) -> None:
-        """Tests that the internal logic from
-        :meth:`ElasticSearch.get_deployment` is correct and filters correctly
-        by test setup value.
-        """
-        mock_query_hits.return_value = self.tests_hits
-
-        jobs_argument = Mock()
-        jobs_argument.value = ['test']
-
-        test_setup = Mock()
-        test_setup.value = ["rpm"]
-
-        jobs = self.es_api.get_deployment(jobs=jobs_argument,
-                                          test_setup=test_setup)
-        self.assertEqual(len(jobs), 1)
-        deployment = jobs['test'].deployment.value
-        network = deployment.network.value
-        self.assertEqual(network.ip_version.value, '4')
-        self.assertEqual(deployment.topology.value, '')
-        test_collection = deployment.test_collection.value
-        self.assertEqual(test_collection.setup.value, 'rpm')
-
-    @patch.object(ElasticSearch, '_ElasticSearch__query_get_hits')
-    def test_get_deployment_test_suites(self, mock_query_hits) -> None:
-        """Tests that the internal logic from
-        :meth:`ElasticSearch.get_deployment` is correct and reads correctly
-        the test suites value.
-        """
-        mock_query_hits.return_value = self.tests_hits
-
-        jobs_argument = Mock()
-        jobs_argument.value = ['test$']
-
-        spec = Mock()
-        spec.value = []
-
-        jobs = self.es_api.get_deployment(jobs=jobs_argument,
-                                          spec=spec)
-        self.assertEqual(len(jobs), 1)
-        deployment = jobs['test'].deployment.value
-        network = deployment.network.value
-        self.assertEqual(network.ip_version.value, '4')
-        self.assertEqual(deployment.topology.value, '')
-        test_collection = deployment.test_collection.value
-        self.assertEqual(test_collection.setup.value, 'rpm')
-        suites = test_collection.tests.value
-        self.assertEqual(suites, set(["neutron", "designate", "octavia"]))
